loans_endpoints/requirements.txt
db-dtypes==1.0.0
numpy==1.21.0
pandas==1.3.0
functions-framework==3.0.0
flask>=1.0,<3.0
google-cloud-bigquery>=1.3.0